moved loadouts to presets
