+++ packages/graph/src/plugins/canvas/themes/presets/light.ts
