+++ packages/graph/src/plugins/canvas/themes/presets/pink.ts
