+++ packages/graph/src/plugins/canvas/themes/presets/shared/node.ts
