+++ packages/graph/src/plugins/canvas/themes/presets/shared/text.ts
